==> 10.read-data.ipynb <==
# Read raw screen data from spreadsheets and reformat

import pandas
pandas.__version__

## Read validated hit compounds

> Of these 72 potential hit compounds, 59 were validated in a secondary screen for dose‐dependency (Appendix Fig S1), with 46 chemicals lengthening and 13 compounds shortening circadian period

hit_df = pandas.read_excel(
    io="data/tamai-screen/emmm201708724-sup-0003-tableev1.xlsx",
    skiprows=1,
)
hit_df.columns = hit_df.columns.str.lower().str.replace(" ", "_")
hit_df = hit_df.dropna(subset=["drug_name"])
hit_df.head(2)

len(hit_df)

## Read FDA compounds

> The SCREENWELL FDA-approved drug library V2 containing 741 compounds was purchased from Enzo Life Sciences (Hayashi Kasei Co., Ltd.), and the International Drug Collection (IDC) containing 311 compounds was purchased from MicroSource Discovery Systems, Inc. (Namiki Shoji Co., Ltd.).

def undo_mangle_dupe_cols(df):
    """
    undo mangle_dupe_cols=True in pandas.read_excel
    Removes ".1" suffix.
    """
    strip_suffix = lambda x: x[:-2] if x.endswith(".1") else x
    df.columns = df.columns.map(strip_suffix)
    return df


def read_fda_spreadsheet(concentration):
    concentration_to_cols = {"1 uM": "B:H", "10 uM": "J:P"}
    renamer = {
        f"Period": "period_before",
        f"{concentration} AVG": "period_after",
        "Period Change": "period_change",
        "HITS/non-HITS": "compound_name",
        "Plate": "plate",
    }
    df = (
        pandas.read_excel(
            io="data/tamai-screen/FDA1_FDA2_Period_24-120.xlsx",
            sheet_name="FDA_1uM_10uM_DRUGS_GRAPHS",
            usecols=f"{concentration_to_cols[concentration]},T",
            mangle_dupe_cols=True,
        )
        .pipe(undo_mangle_dupe_cols)
        .rename(columns=renamer)
        .rename(columns=lambda x: "notes" if x.startswith("Unnamed:") else x)
        # .eval("dead = 0 + (notes == 'dead')")
        .assign(concentration = concentration)
        .dropna(subset=["period_after", "compound_name"])
        .assign(compound_set="fda")
        .convert_dtypes()
    )
    return df


fda_df = pandas.concat([
    read_fda_spreadsheet("1 uM"),
    read_fda_spreadsheet("10 uM"),
])
fda_df.head(2)

## Read IDC compounds

> the International Drug Collection (IDC) containing 311
compounds was purchased from MicroSource Discovery Systems, Inc.
(Namiki Shoji Co., Ltd.).

def read_idc_spreadsheet(concentration):
    concentration_to_cols = {"1 uM": "A:F", "10 uM": "G:L"}
    renamer = {
        "period": "period_before",
        "AVG": "period_after",
        f"IDC {concentration}": "period_change",
        "Name": "compound_name",
    }
    df = (
        pandas.read_excel(
            io="data/tamai-screen/IDC_Period_24-120.xlsx",
            sheet_name="1uM_10uM_Period_Graphs",
            usecols=f"{concentration_to_cols[concentration]},P",
            mangle_dupe_cols=True,
        )
        .pipe(undo_mangle_dupe_cols)
        .rename(columns=renamer)
        .rename(columns=lambda x: "notes" if x.startswith("Unnamed:") else x)
        .assign(concentration = concentration)
        .dropna(subset=["period_after"])
        .assign(compound_set="idc")
        .convert_dtypes()
    )
    return df

idc_df = pandas.concat([
    read_idc_spreadsheet("1 uM"),
    read_idc_spreadsheet("10 uM"),
])
idc_df.head(2)

## Combine compound sets

screen_df = (
    pandas.concat([fda_df, idc_df])
    .convert_dtypes()
    .reindex(columns=[
        "compound_name",
        "compound_set",
        "concentration",
        "period_before",
        "period_change",
        "plate",
        "notes",
    ])
)
screen_df

screen_df.to_csv("data/tamai-screen/combined-results.tsv", sep="\t", index=False, float_format="%.4g")
