==> 12.map-idc-compounds.ipynb <==
# Map IDC Compounds to DrugBank

[MicroSource International Drug Collection](http://www.msdiscovery.com/intdrug.html)

import re
import logging

import pandas
import requests
import requests_cache

requests_cache.install_cache(cache_name="requests-cache", allowable_codes=[200, 404])

## Read IDC compounds from tamai screen spreadsheet

idc_df = pandas.read_excel(
    io="data/tamai-screen/IDC_Period_24-120.xlsx",
    sheet_name="MicroSource Discovery社311品目",
).add_prefix("idc_")
# fix messy column names
idc_df.columns = idc_df.columns.str.strip().str.lower().str.replace(" ", "_")
# http://www.msdiscovery.com/pharmakon.html
idc_df["pharmakon_id"] = idc_df.idc_idnumber.map("Pharmakon1600-{:08d}".format)
idc_df.head(2)


## Map IDC compounds to PubChem and Drugbank

Uses the `Pharmakon1600-` style IDs for the [MicroSource Pharmakon Collection](http://www.msdiscovery.com/pharmakon.html), which includes the IDC. If this fails, fallback to plain name search.

[PubChem](https://pubchempy.readthedocs.io/en/latest/guide/substance.html#retrieving-substances) has the Pharmakon identifiers deposited as substances. For example, [search PubChem](https://pubchem.ncbi.nlm.nih.gov/#query=Pharmakon1600-00300607) for `Pharmakon1600-00300607`.

from utils.pubchem import query_pubchem_for_drugbank

%%time
rows = []
for idc_compound in idc_df.itertuples():
    for compound_name in [idc_compound.pharmakon_id, idc_compound.idc_name]:
        row = query_pubchem_for_drugbank(compound_name)
        if row:
            break
    else:
        logging.warning(f"pharmakon_id and idc_name lookup failed for {idc_compound.idc_name}")
        continue
    row["idc_name"] = idc_compound.idc_name
    row["pharmakon_id"] = idc_compound.pharmakon_id
    rows.append(row)
len(rows)

idc_map_df = (
    idc_df
    [["pharmakon_id", "idc_name", "idc_plate", "idc_position"]]
    .merge(pandas.DataFrame(rows), how="left")
    .convert_dtypes()
)
idc_map_df.head()

# show only records that have drugbank_id
idc_map_df.dropna(subset=["drugbank_id"]).head()

# IDC compounds not mapped to PubChem
idc_map_df[idc_map_df.pubchem_cid.isna()]

print(f"""
{(idc_map_df.query_compound == idc_map_df.pharmakon_id).sum():,} compounds mapped to pubchem via pharmakon_id
{(idc_map_df.query_compound == idc_map_df.idc_name).sum():,} compounds mapped to pubchem via idc_name
{sum(idc_map_df.pubchem_cid.isna()):,} compounds failed to map to pubchem

{sum(idc_map_df.drugbank_id.notna()):,} compounds mapped to DrugBank
""".strip())

idc_map_df.to_csv("data/compounds/idc-to-pubchem.tsv", sep="\t", index=False)

## Map compounds to DrugBank using MyChem.info

Uses https://mychem.info/

Alternative to the PubChem method.

from utils.mychem import query_mychem

idc_df.idc_name[:5]

rows = []
for name in idc_df.idc_name:
    response = query_mychem(name)
    for hit in response.json()["hits"]:
        hit["query_compound"] = name
        rows.append(hit)
query_df = pandas.json_normalize(rows)
query_df.head(2)

query_df = query_df.reindex(columns=["query_compound", "drugbank.id", "drugbank.name", "drugbank.formula", "drugbank.weight.average"])

merged_df = (
    idc_df
    .reindex(columns=["idc_idnumber", "idc_name", "idc_formula", "idc_moecularl_weight"])
    .merge(query_df, left_on="idc_name", right_on="query_compound")
)
merged_df.head(3)

# merged_df.to_csv("data/compounds/idc-to-drugbank.tsv", sep="\t", index=False)

name_mismatch_df = merged_df[merged_df.idc_name.str.lower() != merged_df["drugbank.name"].str.lower()]
name_mismatch_df.head(2)

# mismatched names with different weights
name_mismatch_df[(name_mismatch_df["idc_moecularl_weight"] - name_mismatch_df["drugbank.weight.average"]).abs() > 1.0]
