==> 15.map-fda-compounds.ipynb <==
## Read FDA compounds

https://github.com/greenelab/circadian-clock-modulators/issues/1

> The SCREENWELL FDA-approved drug library V2 containing 741 compounds was purchased from Enzo Life Sciences (Hayashi Kasei Co., Ltd.), and the International Drug Collection (IDC) containing 311 compounds was purchased from MicroSource Discovery Systems, Inc. (Namiki Shoji Co., Ltd.).

import logging

import pandas
import requests_cache

from utils.pubchem import query_pubchem_for_drugbank

requests_cache.install_cache(cache_name="requests-cache", allowable_codes=[200, 404])

### Read tamai screen results

screen_df = pandas.read_table("data/tamai-screen/combined-results.tsv")
screen_df.head(2)

### Read Screen-Well compounds

The spreadsheet contains chemical structures in `mrv` format (ChemAxon MRV format, [Marvin Document](https://docs.chemaxon.com/display/docs/Marvin_Documents_-_MRV.html)). This format appears to be propreitary / not well supported, so map via other fields instead.

# mrv (ChemAxon MRV format)
xlsx_path = "data/compounds/BML-2843_v. 1.4_Rev. 18Dec19_StructureFile.xlsx"
structure_df = pandas.read_excel(xlsx_path, sheet_name=0)
structure_df.head(2)

screenwell_df = pandas.read_excel(xlsx_path, sheet_name=1).add_prefix("screenwell_").convert_dtypes()
# fix messy column names
screenwell_df.columns = screenwell_df.columns.str.strip().str.lower().str.replace(" ", "_")
# remove blank rows
screenwell_df = screenwell_df.query("screenwell_name != 'Blank'")
screenwell_df.head(2)

# Number of compound names in the screen where exact text match fails
screen_names = sorted(screen_df.query("compound_set == 'fda'").compound_name.unique())
len(set(screen_names) - set(screenwell_df.screenwell_name))

%%time
rows = []
for screenwell in screenwell_df.itertuples():
    for by, compound_name in [
            ("name", screenwell.screenwell_iupac),
            ("smiles", screenwell.screenwell_smiles),
            ("name", screenwell.screenwell_name),
        ]:
        row = query_pubchem_for_drugbank(compound_name, by=by)
        if row:
            break
    else:
        logging.warning(f"IUPAC lookup failed for {screenwell.screenwell_name}")
        continue
    row["screenwell_name"] = screenwell.screenwell_name
    rows.append(row)
len(rows)

screenwell_map_df = (
    screenwell_df
    [["screenwell_name", "screenwell_catalog_number", "screenwell_plate_location", "screenwell_smiles", "screenwell_iupac"]]
    .merge(pandas.DataFrame(rows), how="left")
    .convert_dtypes()
)
screenwell_map_df.head(5)

print(f"""
{(screenwell_map_df.query_compound == screenwell_map_df.screenwell_iupac).sum():,} compounds mapped to pubchem via screenwell_iupac
{(screenwell_map_df.query_compound == screenwell_map_df.screenwell_smiles).sum():,} compounds mapped to pubchem via screenwell_smiles
{(screenwell_map_df.query_compound == screenwell_map_df.screenwell_name).sum():,} compounds mapped to pubchem via screenwell_name
{sum(screenwell_map_df.pubchem_cid.isna()):,} compounds failed to map to pubchem

{sum(screenwell_map_df.drugbank_id.notna()):,} compounds mapped to DrugBank
""".strip())

# Screen-Well compounds not mapped to PubChem
screenwell_map_df[screenwell_map_df.pubchem_cid.isna()]

screenwell_map_df.to_csv("data/compounds/screenwell-to-pubchem.tsv", sep="\t", index=False)
